requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=1.0.0
langgraph>=0.2.0
pymilvus>=2.3.0
openai>=1.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
requests>=2.31.0
python-dotenv>=1.0.0
